BOCM-486 - moves service code to own file
