+++ apps/document-storage/package.json
@@ -18,6 +18,7 @@
         "hbs": "^4.1.1",
         "into-stream": "^5.1.1",
         "morgan": "^1.10.0",
-        "multer": "^1.4.1"
+        "multer": "^1.4.1",
+        "crypto-js": "4.1.1"
     }
 }
